+++ NEWS.rst
@@ -4,6 +4,10 @@ News for Kodo C bindings
 This file lists the major changes between versions. For a more detailed list
 of every change, see the Git log.
 
+Latest
+------
+* Minor: Added API for creating debug stacks and printing debug information.
+
 1.4.0
 -----
 * Minor: Added API to control systematic encoding. Systematic encoding means